+++ macros/src/service/ev_fallback_item.rs
@@ -1,5 +1,4 @@
-use super::{kw, Options};
-use crate::doc_string::DocString;
+use super::{kw, ItemOptions, Options};
 use aldrin_codegen::rust::names;
 use proc_macro2::TokenStream;
 use quote::quote;
@@ -8,7 +7,7 @@ use syn::parse::{Parse, ParseStream};
 use syn::{Ident, Result, Token, Type};
 
 pub(super) struct EvFallbackItem {
-    doc: DocString,
+    options: ItemOptions,
     ident: Ident,
     variant: Ident,
     ty: Type,
@@ -38,7 +37,7 @@ impl EvFallbackItem {
     pub(crate) fn gen_variant(&self) -> TokenStream {
         let variant = &self.variant;
         let ty = &self.ty;
-        let doc = &self.doc;
+        let doc = &self.options.doc();
 
         quote! {
             #doc
@@ -50,7 +49,7 @@ impl EvFallbackItem {
     pub(crate) fn layout(&self, options: &Options) -> TokenStream {
         let krate = options.krate();
         let name = self.ident.unraw().to_string();
-        let doc = self.doc.to_introspection();
+        let doc = self.options.doc_alt().to_introspection();
 
         quote! {
             .event_fallback(
@@ -64,7 +63,7 @@ impl EvFallbackItem {
 
 impl Parse for EvFallbackItem {
     fn parse(input: ParseStream) -> Result<Self> {
-        let doc = input.parse()?;
+        let options = input.parse()?;
         input.parse::<kw::event>()?;
         let ident = input.parse::<Ident>()?;
         input.parse::<Token![=]>()?;
@@ -77,7 +76,7 @@ impl Parse for EvFallbackItem {
         );
 
         Ok(Self {
-            doc,
+            options,
             ident,
             variant,
             ty,