+++ macros/src/service/options.rs
@@ -4,6 +4,7 @@ use syn::{parse_quote, Attribute, Error, LitBool, LitStr, Path, Result};
 
 pub(super) struct Options {
     doc: DocString,
+    doc_alt: DocString,
     krate: Path,
     client: bool,
     server: bool,
@@ -15,6 +16,14 @@ pub(super) struct Options {
 impl Options {
     pub(crate) fn doc(&self) -> &DocString {
         &self.doc
+    }
+
+    pub(crate) fn doc_alt(&self) -> &DocString {
+        if self.doc_alt.is_empty() {
+            &self.doc
+        } else {
+            &self.doc_alt
+        }
     }
 
     pub(crate) fn krate(&self) -> &Path {
@@ -47,6 +56,7 @@ impl Parse for Options {
         let attrs = input.call(Attribute::parse_outer)?;
 
         let mut doc = DocString::new();
+        let mut doc_alt = DocString::new();
         let mut krate = parse_quote!(::aldrin);
         let mut client = true;
         let mut server = true;
@@ -56,7 +66,7 @@ impl Parse for Options {
 
         for attr in attrs {
             if attr.path().is_ident("doc") {
-                doc.push(attr)?;
+                doc.push_attr(attr)?;
                 continue;
             }
 
@@ -87,6 +97,9 @@ impl Parse for Options {
                 } else if meta.path.is_ident("schema") {
                     schema = meta.value()?.parse().map(Some)?;
                     Ok(())
+                } else if meta.path.is_ident("doc") {
+                    doc_alt.push(meta.value()?.parse()?);
+                    Ok(())
                 } else {
                     Err(meta.error("unknown attribute"))
                 }
@@ -96,6 +109,7 @@ impl Parse for Options {
         if !introspection || schema.is_some() {
             Ok(Self {
                 doc,
+                doc_alt,
                 krate,
                 client,
                 server,
@@ -106,5 +120,58 @@ impl Parse for Options {
         } else {
             Err(input.error("the attribute `schema` is required to derive Introspectable"))
         }
+    }
+}
+
+pub(super) struct ItemOptions {
+    doc: DocString,
+    doc_alt: DocString,
+}
+
+impl ItemOptions {
+    pub(crate) fn doc(&self) -> &DocString {
+        &self.doc
+    }
+
+    pub(crate) fn doc_alt(&self) -> &DocString {
+        if self.doc_alt.is_empty() {
+            &self.doc
+        } else {
+            &self.doc_alt
+        }
+    }
+}
+
+impl Parse for ItemOptions {
+    fn parse(input: ParseStream) -> Result<Self> {
+        let attrs = input.call(Attribute::parse_outer)?;
+
+        let mut doc = DocString::new();
+        let mut doc_alt = DocString::new();
+
+        for attr in attrs {
+            if attr.path().is_ident("doc") {
+                doc.push_attr(attr)?;
+                continue;
+            }
+
+            if !attr.path().is_ident("aldrin") {
+                return Err(Error::new_spanned(
+                    attr,
+                    "extected attributes `aldrin` or `doc`",
+                ));
+            }
+
+            attr.parse_nested_meta(|meta| {
+                if meta.path.is_ident("doc") {
+                    doc_alt.push(meta.value()?.parse()?);
+                    Ok(())
+                } else {
+                    Err(meta.error("unknown attribute"))
+                }
+            })?;
+        }
+
+        Ok(Self { doc, doc_alt })
     }
 }