+++ macros/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `service!()` now supports doc comments and also sets them on the service's introspection.
 - The `RefType` derive macro now supports doc comments.
 - `Introspectable` now also supports doc string.
+- Add `#[aldrin(doc = "...")]` attribute to provide alternative documentation for introspection.
 
 ### Changed